feat: GameRoom 도입으로 다중 방 독립 게임 구조 구현 (#6)
